Add files via upload
From Drought to Floods: A Step-by-Step Guide

This project was carried out through the Impact Seed Fund (ISF), an initiative by the Pulitzer Center.

The same analysis that we performed in this guide for Ciudad del Plata could easily be replicated using open datasets from your own country. By following the step-by-step process laid out here and adapting it to locally available geospatial, environmental, and census data, you can identify and analyze vulnerable areas in the face of climate change.

Warning
This guide is designed to help you understand and use some basic functions of Geographic Information System (GIS) software. The example case presented is for instructional purposes only. The final results obtained from this guide should be considered as an approximation to the topic of study and not as definitive or corroborated conclusions. To properly interpret the results and apply the knowledge effectively, it is highly recommended to involve subject matter experts. We emphasize the importance of this recommendation to ensure rigorous and accurate analysis.